+++ backend/api/serializers/__init__.py
@@ -3,8 +3,6 @@ from .computeplan import ComputePlanSerializer
 from .computetask import ComputeTaskInputAssetSerializer
 from .computetask import ComputeTaskOutputAssetSerializer
 from .computetask import ComputeTaskSerializer
-from .computetask import LegacyComputeTaskSerializer
-from .computetask import LegacyComputeTaskWithRelationshipsSerializer
 from .datamanager import DataManagerSerializer
 from .datamanager import DataManagerWithRelationsSerializer
 from .datasample import DataSampleSerializer
@@ -21,8 +19,6 @@ __all__ = [
     "ComputeTaskSerializer",
     "ComputeTaskInputAssetSerializer",
     "ComputeTaskOutputAssetSerializer",
-    "LegacyComputeTaskSerializer",
-    "LegacyComputeTaskWithRelationshipsSerializer",
     "DataManagerSerializer",
     "DataManagerWithRelationsSerializer",
     "DataSampleSerializer",