+++ CHANGELOG.md
@@ -25,10 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Algo creation events aren't included in newsfeed anymore.
-
-### Changed
-
 - Remove task category from the compute engine.
+- BREAKING: compute task specific endpoints.
+- BREAKING: category related fields to create task.
 
 ## [0.32.0] 2022-10-03